enlarge graph for flux diagram round a central species
instead of drawing only the species in the 1st layer round a central
species, the method is updated to draw a user-defined of species
round a central species by systematically enlarging the graph with
"neighbouring" species.

This is done in a recursive way.

For now, a fixed number of species is
contained in the graph (10). Later on, this could be passed on as an
argument.

TODO: graph layout could be drawn radially (e.g. using twopi) with the
central species in the middle, with concentric circles of neighbouring
species.

Tested successfully, yet no unit test. I don't expect many risks of break-
ing other stuff since a separate method is created and the if...else
clause for centralSpecies in generateFluxDiagram is preserved.