generateCentralSpeciesFluxSnapshot methods added
much alike the generateFluxDiagram created a video animation of
the reaction system around a central species, this method now
does the same, but adds labels next to the edges that represent
the share of a particular edge (= flux between two species) in the
total flux of consumption or formation of a particular central species.

The output is a number of snapshots (.png's) stored in a user-defined
directory, yet not stitched together in an animation.

Basically, both createFluxDiagram and generateFluxDiagram are tweaked and
copied to the new methods.

I chose not to interrupt the existing methods with optional label edges
representing the share of the total flux but to create new methods that
eventually could be migrated to a module with features for automated
quantitative rate-of-production analysis.