proof of concept of the error-cancelling reaction generation module
The idea is to provide a molecule, and this module will generate the
products and reactants that go with the given molecule according
to the error-cancellation scheme of the Connectivity-Based Hierarchy method.

Implemented the first rung (cbh-0) as a class CBH0Reaction that
extends an (abstract) CBH_X_Reaction superclass with common
attributes and methods for the remainder of the rungs (cbh-1 till cbh-3).

Added a test module