redesign of error_cancelling.py module
use of classes to structure the code more:

1) Parser classes (superclass: AbstractErrorCancellingReactionParser)
that read in multiple types of input files for the
error cancelling estimator.

Current implementations:
a) UserDefinedReactionParser: the user provides an error cancelling
reaction and the code will use that reaction to reduce the error
on the enthalpy of the 'unknown' species.
b) UserDefinedSpeciesReactionParser: the implementation generates
a reaction according to the Connectivity-based Hierarchy of
the Raghavachari group.

2) ErrorCalculator: this class carries out the calculations for
thermo and so forth.