correct hash function of reaction.Reaction
reactants/products are lists and preserve order. The hash of this list
would be different when the order of two reactant molecules would
be reversed, and would lead to different hashes, and errors in Set's.

It is therefore opted to calculate the sum of the hashes of the reactants,
which will reduce the entropy of the hash, but avoid erroneous Sets.