move the template matching to when we generate kinetics
because we now have a persisten way of storing the atom labels
on the level of the TemplateReaction we can move the matching
of the reactive centers to template Groups from __generateReactions
to the point where we want to generate the Kinetics of a reaction.

However, because the atom labels are referring to the first molecule
in the list of resonance isomers of a Species, we cannot shuffle them
around anymore. Therefore, we remove the re-ordering of resonance
isomer based on the thermochemistry. The thermo is already stored
on the species level, which prevents us from using the wrong thermo.