create fromOBMol method in Molecule
does the same as fromRDKitMol, but from openbabel. Because OB does
not have a mol.GetAtomWithIdx(i) method, two independent loops are
created. One iterating over the atoms and adding them to self.vertices,
one loop iterating over obmol's bonds and using the idx of the atom
as an index in the self.vertices array.

- Does NOT kekulize the structure, so specific bonds might be aromatic.