replace reaction param by reactants in groups.KinGr.getReactionTemplate
this allows us to use TemplateReaction.reactants as a list of
augmented inchi's rather than a list of Species.

We convert the list of augmented inchi's to a list of Species that
we pass to the self.groups.getReactionTemplate(reactants) method.

For the other calls to self.groups.getReactionTemplate(reactants),
the reactants are stored as Species already.