inflate the products when they are used for the reverse reaction
in KineticsFamily.generateReactions(...) we use the products
of previously generated reactions to generate new reactions for
reaction families that are their own reverse.

Because the products of a TemplateReaction are now simply
augmented inchi's, we need to convert the aug. inchi's back to graphs
so that they can undergo graph transformations.