only sort atoms in products in __createReaction
We can assume that the atoms in the reactants are already sorted. Also,
sorting the atoms after the reactant mapping has been carried out,
poses a risk if the mapping relies on the order of the atoms. In the
current implementation we use atom references, but in the future this
might change.